requirements.txt
social-auth-app-django
kiwitcms-tenants
